requirements.txt
scipy
numpy
numericalunits
matplotlib
svgwrite
requests
pycoordinates
attrs